+++ apps/shared-tree-shaking/no-server/host/CHANGELOG.md
@@ -1,5 +1,23 @@
 # modernjs-ssr-nested-remote
 
+## 1.0.19
+
+### Patch Changes
+
+- @module-federation/enhanced@2.9.0
+
+## 1.0.18
+
+### Patch Changes
+
+- @module-federation/enhanced@2.8.2
+
+## 1.0.17
+
+### Patch Changes
+
+- @module-federation/enhanced@2.8.1
+
 ## 1.0.16
 
 ### Patch Changes