+++ apps/shared-tree-shaking/with-server/provider/CHANGELOG.md
@@ -1,5 +1,24 @@
 # modernjs-ssr-nested-remote
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [e68f1b5]
+  - @module-federation/modern-js-v3@2.9.0
+
+## 1.0.18
+
+### Patch Changes
+
+- @module-federation/modern-js-v3@2.8.2
+
+## 1.0.17
+
+### Patch Changes
+
+- @module-federation/modern-js-v3@2.8.1
+
 ## 1.0.16
 
 ### Patch Changes